move evaluate_ functions to interpreter.c
